move data types
