+++ components/home/all-projects.tsx
@@ -3,7 +3,7 @@
 import { useState } from "react"
 import Link from "next/link"
 import Image from "next/image"
-import { projects } from "../../lib/data"
+import { projects } from "../../data/sampleProjects"
 import { Badge } from "..//ui/badge"
 import { Button } from "../ui/button"
 import { ChevronLeft, ChevronRight } from "lucide-react"
