+++ components/home/featured-projects.tsx
@@ -4,7 +4,7 @@ import { useState, useRef } from "react"
 import Link from "next/link"
 import Image from "next/image"
 import { ChevronLeft, ChevronRight } from "lucide-react"
-import { projects } from "../../lib/data"
+import { projects } from "../../data/sampleProjects"
 import { Badge } from "../ui/badge"
 import { Button } from "../ui/button"
 import { cn } from "../../lib/utils"
